moved composite under reference, removed the check from penalty_sat function
